Refactor export module tests and rename view.rs
Moved and consolidated tests from integration_test.rs and test.rs into collectors.rs and tensor_view.rs, respectively. Renamed view.rs to tensor_view.rs and updated module exports to reflect the new file structure. This improves test organization and code clarity.
