+++ crates/burn-core/src/module/export/mod.rs
@@ -1,12 +1,7 @@
 mod collectors;
-mod view;
-
-#[cfg(test)]
-mod integration_test;
-#[cfg(test)]
-mod test;
 mod module_ext;
+mod tensor_view;
 
 pub use collectors::TensorViewCollector;
-pub use view::TensorView;
 pub use module_ext::ModuleExport;
+pub use tensor_view::TensorView;
