+++ crates/burn-core/src/module/export/tensor_view.rs
@@ -37,3 +37,43 @@ impl TensorView {
         (self.data_fn)()
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::TestBackend;
+
+    #[test]
+    fn test_tensor_view_float() {
+        let device = Default::default();
+        let tensor = Tensor::<TestBackend, 2>::from_data([[1.0, 2.0], [3.0, 4.0]], &device);
+
+        let view = TensorView::from_float(&tensor);
+        let data = view.to_data();
+
+        assert_eq!(data.shape, vec![2, 2]);
+    }
+
+    #[test]
+    fn test_tensor_view_int() {
+        let device = Default::default();
+        let tensor = Tensor::<TestBackend, 2, Int>::from_data([[1, 2], [3, 4]], &device);
+
+        let view = TensorView::from_int(&tensor);
+        let data = view.to_data();
+
+        assert_eq!(data.shape, vec![2, 2]);
+    }
+
+    #[test]
+    fn test_tensor_view_bool() {
+        let device = Default::default();
+        let tensor =
+            Tensor::<TestBackend, 2, Bool>::from_data([[true, false], [false, true]], &device);
+
+        let view = TensorView::from_bool(&tensor);
+        let data = view.to_data();
+
+        assert_eq!(data.shape, vec![2, 2]);
+    }
+}
